Added Camera Scrolling and Background
By pressing C, you can enable/disable camera scrolling.  It'll attach to
wherever you character currently stands and move with your character.

It's a bit buggy when you reach the edge of the map.

Also added in a very large version of the ALTTP map, which is now the
background.