Created a Map class, Movement is relative to Map
Slightly modified how the main loop proceeds.  It now calls render on
the Map object, and update on the Map object.  The Map object then calls
render/update on the Characters within itself.

There are some bugs with rendering Chars on the left and bottom parts of
the screen when they are on the edge.

There may be a bug with NPCs on the edge of the map (uncertain)

Bug: The player is not completely centered.